requirements.txt
absl-py==2.1.0
certifi==2024.8.30
charset-normalizer==3.3.2
cloudpickle==1.6.0
contourpy==1.2.1
cycler==0.12.1
Farama-Notifications==0.0.4
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
future==1.0.0
grpcio==1.66.1
gym==0.17.3
gym-chess==0.1.1
gym-notices==0.0.8
gymnasium==0.29.1
idna==3.10
importlib_metadata==8.0.0
importlib_resources==6.4.0
Jinja2==3.1.4
kiwisolver==1.4.5
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.9.1
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
packaging==24.1
pandas==2.2.2
pillow==10.4.0
protobuf==5.28.0
pyglet==1.5.0
pyparsing==3.1.2
python-chess==0.31.4
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
scipy==1.13.1
six==1.16.0
stable-baselines3==1.1.0
sympy==1.12.1
tensorboard==2.17.1
tensorboard-data-server==0.7.2
torch==2.3.1
triton==2.3.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
Werkzeug==3.0.4
zipp==3.19.2
